Alter `dbm-gorm-diff` command
Optionally stop rendering of addDefaultValue changes using the new configuration property
Do not render any changes which are not in the default schema.
This prevents any gorm domains which have non-default schema schemas from being rendered into the changeset.
This is required as at the moment the changeset does not render schema attributes, as such any tables not in the provided default schema are added to the changeset, which is not desired.